backend/requirements.txt
amadeus==12.0.0
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
django==5.2.10
django-prometheus==2.4.1
djangorestframework==3.16.1
gunicorn==24.1.0
idna==3.11
iniconfig==2.3.0
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
prometheus-client==0.24.1
psycopg==3.3.2
pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
